Switch from yargs to commander
The problems which previously motivated me to switch from commander to
yargs have been solved (by the addition of .exitOverride() and
.configureOutput()).  Commander has more flexibility than yargs for
option processing using .on('option'), which can be used for options
which are sensitive to ordering or have more complicated handling.  It
is also much simpler and has less exotic behavior that needs to be
disabled (see all the yargs options which were set).  Finally, it is
about 1/6 the total size.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>