Use Symbol options to test cli
To test cli, have it call a function passed as a private Symbol on the
options object instead of modulename.func, when provided.

Note: The Symbol is defined in lib/symbols.js, which is not included in
package.json#exports, so can not be imported outside of this package and
does not increase the API surface of the module.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>