re-throw non-Commander parse errors
Previously, if an error was thrown by command.parse, it was assumed to
have been written to stderr.  However, that was not true for errors
other than the ones expected by Commander (e.g. InvalidArgumentError),
which would cause the command to exit without printing an error.  This
is terrible UX, since all the user knows is that the command failed
(exit code 1) but not why.

CLI parsing code should throw errors derived from
commander.CommanderError (e.g. commander.InvalidArgumentError).  Treat
any other error as unhandled and re-throw it (to cause
unhandledRejection and likely print a backtrace then exit).  This way
the user is informed of the error and the developer has information to
fix it.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>